+++ srcs/core/docker-compose.yml
@@ -6,7 +6,7 @@
 #    By: fclivaz <fclivaz@student.42lausanne.ch>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/05/04 22:27:45 by fclivaz           #+#    #+#              #
-#    Updated: 2025/05/26 22:36:35 by fclivaz          ###   LAUSANNE.ch        #
+#    Updated: 2025/06/19 18:12:11 by fclivaz          ###   LAUSANNE.ch        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -31,6 +31,3 @@ services:
     environment:
       - API_KEY
       - TZ=Europe/Zurich
-    depends_on:
-      - sarif_db
-    # restart: on-failure
