+++ srcs/database/package.json
@@ -10,6 +10,8 @@
   "author": "fclivaz",
   "license": "GPL-3.0",
   "dependencies": {
+    "@fastify/multipart": "^9.0.3",
+    "@fastify/static": "^8.2.0",
     "axios": "^1.9.0",
     "better-sqlite3": "^11.9.0",
     "fastify": "^5.3.2"
