+++ srcs/database/srcs/db_extras.ts
@@ -6,7 +6,7 @@
 //   By: fclivaz <fclivaz@student.42lausanne.ch>    +#+  +:+       +#+        //
 //                                                +#+#+#+#+#+   +#+           //
 //   Created: 2025/06/18 20:58:12 by fclivaz           #+#    #+#             //
-//   Updated: 2025/06/19 00:05:44 by fclivaz          ###   LAUSANNE.ch       //
+//   Updated: 2025/06/19 21:36:16 by fclivaz          ###   LAUSANNE.ch       //
 //                                                                            //
 // ************************************************************************** //
 
@@ -14,9 +14,13 @@ import DatabaseWorker from "./db_methods.ts"
 import Axios from "axios"
 import type * as at from "axios"
 import type * as fft from 'fastify'
+import * as fs from "node:fs"
+import { pipeline } from "node:stream"
+import util from 'util';
 import { randomBytes, scrypt } from "crypto"
+import type { MultipartFile } from "@fastify/multipart"
 
-export async function check_password(user: string, field: "DisplayName" | "EmailAddress", password: string) {
+async function check_password(user: string, field: "PlayerID" | "DisplayName" | "EmailAddress", password: string) {
 	return new Promise((resolve, reject) => {
 		let user_data: object = {};
 		try {
@@ -44,13 +48,38 @@ export async function check_password(user: string, field: "DisplayName" | "Email
 interface Disname {
 	DisplayName: string
 }
-export async function passck_route(fastify: fft.FastifyInstance, options: fft.FastifyPluginOptions) {
+
+const pump = util.promisify(pipeline);
+
+export async function extra_routes(fastify: fft.FastifyInstance, options: fft.FastifyPluginOptions) {
+
 	fastify.get<{ Params: Disname }>("/check/:DisplayName", async function handler(request: fft.FastifyRequest, reply: fft.FastifyReply) {
 		const parms: Disname = request.params as Disname;
 		const headrs = request.headers as object;
 		if (await check_password(parms.DisplayName, "DisplayName", headrs["password"]))
-			reply.code(200).send({ skill: "solution" })
+			return reply.code(200).send({ skill: "solution" })
 		else
-			reply.code(403).send({ skill: "issue" })
+			return reply.code(403).send({ skill: "issue" })
+	})
+
+	fastify.get<{ Params: Disname }>("/download/:DisplayName", async function handler(request: fft.FastifyRequest, reply: fft.FastifyReply) {
+		const parms: Disname = request.params as Disname;
+		if (fs.existsSync(`${process.env.FILELOCATION}/${parms.DisplayName}.png`))
+			return reply.code(200).sendFile(`${parms.DisplayName}.png`)
+		else
+			return reply.code(404).send({ skill: "issue" })
+	})
+
+	fastify.post<{ Params: Disname }>("/upload/:DisplayName", async function handler(request: fft.FastifyRequest, reply: fft.FastifyReply) {
+		const parms: Disname = request.params as Disname;
+		const part = await request.file({
+			limits: {
+				fileSize: 4 * 1024 * 1024
+			}
+		}) as MultipartFile
+
+		const filename = parms.DisplayName + part?.filename.slice(part?.filename.lastIndexOf('.'))
+		await pump(part.file, fs.createWriteStream(`${process.env.FILELOCATION}/${filename}`));
+		return reply.code(200).send({ skill: "solution" })
 	})
 }
