+++ srcs/database/srcs/db_startup.ts
@@ -1,0 +1,126 @@
+// ************************************************************************** //
+//                                                                            //
+//                                                        :::      ::::::::   //
+//   db_startup.ts                                      :+:      :+:    :+:   //
+//                                                    +:+ +:+         +:+     //
+//   By: fclivaz <fclivaz@student.42lausanne.ch>    +#+  +:+       +#+        //
+//                                                +#+#+#+#+#+   +#+           //
+//   Created: 2025/03/18 17:42:46 by fclivaz           #+#    #+#             //
+//   Updated: 2025/06/19 19:16:40 by fclivaz          ###   LAUSANNE.ch       //
+//                                                                            //
+// ************************************************************************** //
+
+import Database from "better-sqlite3"
+import { tables } from "./db_vars.ts"
+import type * as sqlt from "better-sqlite3"
+import { hash_password } from "./db_helpers.ts"
+import DatabaseWorker from "./db_methods.ts"
+
+//
+// Automatically check if any column has been changed in db_vars.ts
+// If changes are detected, rows are dropped or added automatically using the arguments given.
+// NOTE: Argument changes do NOT get reflected yet!
+//
+
+function check_db_columns(db: sqlt.Database, table: string) {
+	const result = db.pragma(`table_info(${table})`) as Array<object>
+	const dbrows: Array<string> = [];
+	for (const item of result) {
+		dbrows.push(item["name"]);
+		if (tables[table]["Fields"].indexOf(item["name"]) === -1) {
+			db.prepare(`ALTER TABLE ${table} DROP COLUMN ${item["name"]}`).run()
+			console.log(`DB INIT: dropped ${item["name"]}.`)
+		}
+	}
+	for (const item of tables[table]["Fields"]) {
+		if (dbrows.indexOf(item) === -1 && item.length > 0) {
+			db.prepare(`ALTER TABLE ${table} ADD ${item} ${tables[table]["Arguments"][tables[table]["Fields"].indexOf(item)]}`).run()
+			console.log(`DB INIT: added ${item}.`)
+		}
+	}
+}
+
+//
+// Automatically and asynchronously initialize the database.
+// Creates the necessary tables, fields and arguments using db_vars.ts
+// Procedurally generates the SQL, so any modification in db_vars.ts will get reflected onto the live database.
+// Returns a boolean Promise. True if the database was succesfully initialized, False if not.
+//
+
+export async function init_db() {
+	return new Promise((resolve, reject) => {
+		const db = new Database(process.env.DBLOCATION)
+		try {
+			for (const item in tables) {
+				const result = db.pragma(`table_info(${tables[item].Name})`) as Array<object>
+				if (result.length === 0) {
+					if (tables[item].Fields.length !== tables[item].Arguments.length) {
+						console.warn(`${tables[item].Name}'s Fields and Arguments are of different lengths. Skipping.`)
+						continue;
+					}
+					let sql = `CREATE TABLE ${tables[item].Name} (\n`;
+					for (let x = 0; x < tables[item].Fields.length; x++) {
+						if (x === tables[item].Fields.length - 1)
+							sql += `${tables[item].Fields[x]} ${tables[item].Arguments[x]})`
+						else
+							sql += `${tables[item].Fields[x]} ${tables[item].Arguments[x]},\n`
+					}
+					db.prepare(sql).run()
+					console.log(`DB INIT: Created table ${tables[item].Name}`);
+				}
+				else
+					check_db_columns(db, tables[item].Name);
+			}
+			let sql = `INSERT INTO ${tables.Players.Name} VALUES (`;
+			const body = {
+				PlayerID: "P-0",
+				DisplayName: "Guest",
+				EmailAddress: "null",
+				Password: "0"
+			}
+			for (const key of tables.Players.Fields) {
+				if (key !== "")
+					sql += ` @${key},`
+				if (body[key] === undefined)
+					body[key] = null
+			}
+			db.prepare(sql.replace(/.$/, ")")).run(body)
+		} catch (error) {
+			console.error(error);
+			reject(false);
+		} finally {
+			db.close()
+		}
+		resolve(true);
+	})
+}
+
+//
+// Adds a default user with Admin privileges and guest data.
+//
+
+export async function add_default_user() {
+	return new Promise((resolve, reject) => {
+		(async () => {
+			try {
+				const test = DatabaseWorker.get(tables.Players.Name, tables.Players.Fields[1], process.env.ADMIN_NAME as string)
+				if (test[tables.Players.Fields[1]] === process.env.ADMIN_NAME as string)
+					resolve(true);
+			} catch (e) {
+				if (e.code !== 404) {
+					console.dir(e);
+					reject(false);
+				}
+			}
+			const default_user: object = { DisplayName: process.env.ADMIN_NAME, EmailAddress: "DEFAULTEMAIL", Admin: 1 };
+			try {
+				default_user["Password"] = await hash_password(process.env.ADMIN_PASSWORD as string)
+				DatabaseWorker.post(tables.Players.Name, tables.Players.Fields, default_user)
+			} catch (e) {
+				console.dir(e);
+				reject(false);
+			}
+			resolve(true);
+		})()
+	})
+}
